feat(api): バージョン情報 API と更新履歴キャプションを追加
Vite プラグイン（vite.config.ts inline）で 2 つの JSON エンドポイントを提供。
dev サーバー・preview サーバー・本番ビルド dist/ の 3 経路で同じ内容を配信する。

- GET /api/version(.json) → { "version": "0.15" }
- GET /api/version/detail(.json) → { "revisions": [...] }

本番ビルドでは dist/api/version.json と dist/api/version/detail.json を静的に出力するため、
静的ホスティング時も追加設定不要で利用できる。

UI とプラグインの双方から参照する単一ソースを src/data/revisions.ts に切り出し、
既存の import 経路との互換のため RevisionHistory.tsx から再エクスポートを維持。

ヘッダーのグレーキャプションに History アイコン + 「更新履歴」ラベルを追加し、
何の情報かを一目で判別できるように改善。

- npx tsc --noEmit クリーン
- npx vite build 成功、dist/api/version.json + dist/api/version/detail.json を出力
- 手元 curl で /api/version, /api/version.json, /api/version/detail, /api/version/detail.json を確認
- npm run test:e2e:ci で 98/98 pass（回帰なし）
